remove all dependencies to misty
