+++ app/controllers/dashboard_controller.rb
@@ -98,8 +98,7 @@ class DashboardController < ::ScopeController
     )
   end
 
-  rescue_from 'Core::Api::Error',
-              'Elektron::Errors::ApiResponse' do |exception|
+  rescue_from 'Elektron::Errors::ApiResponse' do |exception|
     options = {
       title: exception.code_type, description: :message,
       warning: true, sentry: false
