--- lib/core/api.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'api/error'
-require_relative 'api/client_wrapper'
-require_relative 'api/client_manager'
-
-module Core
-  # This module contains all classes needed for the communication with the API.
-  module Api
-  end
-end
